Change player to be just a struct...
This is necessary because we want PlayerList to update
when the Player is changed in PlayerDetail. PlayerList
only listens to an array of Players, in ListData. This is
fine, except that when Player is a class, changes to the
properties of a player element do not cause ListData to emit
a notification to PlayerList that it should update. By changing
Player to a struct, any changes to its properties will change the
array and thus cause PlayerList to update.